refactore initContainerd
